apps/dj-backend/requirements.txt
-r requirements-production.txt

#
markdown  # django browseable api (REST FRAMEWORK)

# testing
mock
model_bakery

pre-commit

# linters, formatters, ...
Pylint          # linter
pylint-mccabe
pylint-django
black           # reformatter
